src/analyzers/image_analyzer.py
import cv2
import logging
import io

from PIL import Image
from PIL.ExifTags import TAGS

logger = logging.getLogger(__name__)


class ImageAnalyzer:
    """Analyzes images for AI-generation indicators."""

    async def analyze_image(self, image_path: str):
        """Perform comprehensive image analysis."""
        # 1. Check EXIF Data
        img = Image.open(io.BytesIO(image_path))
        exif_data = {TAGS.get(k): v for k, v in img._getexif().items() if k in TAGS} if img._getexif() else {}
        
        # 2. Basic Manipulation Detection (Error Level Analysis)
        # Check for inconsistent compression levels using OpenCV
        cv_img = cv2.imread(image_path)
        # (Simplified logic for tutorial purposes)
        return {"exif": exif_data, "status": "Analyzed locally"}