pyproject.toml
[project]
name = "phishing-detector"
version = "0.1.0"
description = "AI-powered phishing and deepfake detection system"
requires-python = ">=3.11"
authors = [
    {name = "Nandita Sreekumar", email = "sreekumar.nandita1@gmail.com"}
]
license = {text= "MIT"}

dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "ollama>=0.1.0",
    "httpx>=0.27.0",
    "python-whois>=0.9.0",
    "dnspython>=2.6.0",
    "Pillow>=10.2.0",
    "opencv-python-headless>=4.9.0",
    "aiosqlite>=0.19.0",
    "streamlit>=1.31.0",
    "plotly>=5.18.0",
    "python-multipart>=0.0.9",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"